Refactor parameter search scripts
