+++ finetuning/v2/evaluation/evaluate_automatic_segmentation.py
@@ -4,13 +4,11 @@ Two modes share one pipeline:
   ais: the jointly finetuned UniSAM2 decoder, turned into instances by the flow / multicut postprocessing.
   apg: the same decoder proposes candidates, which the interactive branch turns into masks.
 
-One invocation tunes on the validation split and then scores the test split with the parameters it
-selected. A dataset with no data held out from the evaluation keeps the library defaults, because a
+The parameters are tuned separately, once per dataset and mode, by `parameter_search.py`. This
+script only reads what that sweep found (see `common.read_tuned_params`) and scores the test split
+with it. A dataset with no data held out from the evaluation keeps the library defaults, because a
 sweep there would select its parameters on the very samples the reported score is measured on. See
 `common.VAL_SPLITS` for what counts as held out.
-
-EM datasets are tuned in dense (multicut) mode and ranked by the CREMI score, all others in sparse
-(flow) mode and ranked by mSA.
 
 Usage examples:
     python evaluate_automatic_segmentation.py -d livecell -m hvit_b -e <exp> --mode ais
@@ -19,433 +17,21 @@ Usage examples:
 """
 
 import os
-import time
 import json
 import argparse
 import warnings
-import itertools
-import threading
-from concurrent import futures
 
-import numpy as np
 import pandas as pd
 from tqdm import tqdm
 
 import torch
 
-from elf.evaluation import mean_segmentation_accuracy
-
-from bioimage_cpp.segmentation import label as connected_components, watershed
-
-from micro_sam.v2.postprocessing import watershed_heightmap, _compute_flow_density
-
 from common import (
-    DATA_ROOT, DATASETS_2D, DATASETS_3D, DATASETS_3D_EM, DATASET_SPACING, VAL_Z_RANGE,
-    GT_MIN_SIZE_2D, CROP_SHAPE_3D, VOLUME_SPEED_OPTIONS, MODEL_TYPES,
-    build_apg_segmenter, check_data_download, checkpoint_checksum, combine_checkpoint_checksums,
-    drop_severed_objects, export_joint_checkpoint, genuine_misses, get_joint_checkpoint, has_val_split,
-    load_data, load_unisam2_model, n_samples, postprocess_unisam2, predict_unisam2, read_tuned_params,
-    run_dataset_evaluation,
+    DATA_ROOT, DATASETS_2D, DATASETS_3D, DATASET_SPACING, GT_MIN_SIZE_2D, MODEL_TYPES, MODES,
+    VOLUME_SPEED_OPTIONS, build_model, check_data_download, drop_severed_objects, genuine_misses,
+    has_val_split, load_data, n_samples, postprocess_unisam2, predict_unisam2, read_tuned_params,
+    resolve_checkpoint_identity, run_dataset_evaluation,
 )
-
-MODES = ("ais", "apg")
-
-# Sparse (flow) grid for LM data. Keys map to flow_instance_segmentation arguments.
-LM_GRID = {
-    "foreground_threshold": [0.3, 0.4, 0.5, 0.6, 0.7],
-    "density_threshold": [5.0, 10.0, 20.0],
-    "min_size": [10, 25, 50, 100, 200, 500],
-    "sigma": [0.25, 0.5, 1.0, 2.0],
-    "n_iter": [50, 100, 200],
-    "dt": [0.25, 0.5, 1.0],
-    "foreground_weight": [0.0, 0.25, 0.5, 0.65, 0.75, 0.9],
-}
-
-# Dense (multicut) grid for EM data. Keys map to run_multicut arguments.
-EM_GRID = {
-    "beta": [0.5, 0.6, 0.7, 0.8],
-    "density_threshold": [3.0, 5.0, 10.0],
-    "sigma": [0.5, 1.0, 2.0],
-    "n_iter": [25, 50],
-}
-
-# Keys map to AutomaticPromptGenerator.generate. Only 'candidate_threshold' and 'sigma' re-prompt
-# SAM2; the rest select among proposals that are already there, so they are swept densely for free.
-# See APG_PROPOSAL_KEYS and score_sample_apg_cached. 'foreground_threshold' and 'min_size' are pinned
-# to their defaults: each moved the result by less than 0.001, which does not pay for a larger sweep.
-APG_GRID_2D = {
-    "foreground_threshold": [0.7],
-    "candidate_threshold": [1.0, 1.5, 2.25, 3.0],
-    "n_iter": [50],
-    "sigma": [0.5, 1.0, 2.0],
-    "min_candidate_size": [1, 4],
-    "score_threshold": [0.5, 0.6, 0.7, 0.8],
-    "max_overlap": [0.15, 0.3],
-    "min_size": [50],
-    "refine_with_box_prompts": [False, True],
-}
-
-# A volume gates its propagation on the score, so none of its axes are free and this grid stays small.
-# A candidate's convergence density scales with the object's size, so small nuclei never clear a single
-# rung of the ladder: hence the lower bottom rungs, less smoothing and a smaller component floor.
-# 'max_overlap' is swept here and pinned in the 2d grid, because of the failure it causes on a volume:
-# a neighbour clips a fifth of an object, the object's own well-propagated mask is then rejected
-# wholesale, and the object is left unclaimed.
-APG_GRID_3D = {
-    "candidate_threshold": [(1.5, 10.0), (1.0, 5.0), (0.5, 5.0), (0.25, 1.0, 5.0)],
-    "sigma": [0.25, 0.5, 1.0],
-    "min_candidate_size": [1, 4],
-    "score_threshold": [0.6],
-    "max_overlap": [0.15, 0.5],
-    "min_size": [100],
-}
-
-# The APG parameters that decide the mask proposals, which is the half that needs the model. Combos
-# that share these reuse one set of proposals, as the sparse grid reuses a cached flow density.
-APG_PROPOSAL_KEYS = ("candidate_threshold", "sigma", "foreground_threshold", "n_iter", "dt", "min_candidate_size")
-
-# Sparse-flow parameters that determine the (expensive) convergence-density map. Combos that share
-# these reuse a cached density. Only 'density_threshold', 'foreground_weight' and 'min_size' (the cheap
-# seed + watershed + size-filter steps) then vary on top. See score_image_sparse_cached.
-FLOW_DENSITY_KEYS = ("foreground_threshold", "sigma", "n_iter", "dt")
-
-# Dense-multicut parameters that determine the (expensive) slice-wise oversegmentation and RAG.
-# Combos that share these reuse a cached oversegmentation. Only 'beta' (the cheap edge-cost +
-# multicut-solve step) then varies on top. See score_image_dense_cached.
-OVERSEG_KEYS = ("density_threshold", "sigma", "n_iter", "dt")
-
-# Metric the grid is ranked by, and whether lower is better. 'msa' (mean segmentation accuracy) is
-# maximised; 'cremi' (a VI + adapted-Rand combination for neuron segmentation) is minimised.
-CRITERION_ASCENDING = {"msa": False, "cremi": True}
-
-POSTPROC_THREADS = 4
-
-
-def tuning_config(dataset_name, mode, criterion=None, crop_shape=None):
-    """Build the sweep configuration for one dataset and mode.
-
-    EM datasets are tuned in dense (multicut) mode and ranked by the CREMI score, all others in
-    sparse (flow) mode and ranked by mSA. With mode='apg' the postprocessing is replaced by the
-    prompt generation, which is swept over its own grid; the ranking metric follows the data either
-    way, so an APG result is directly comparable with the AIS result of the same dataset.
-
-    Args:
-        dataset_name: The dataset to tune on.
-        mode: The segmentation mode, 'ais' or 'apg'.
-        criterion: The metric the grid is ranked by. Defaults to the data-specific choice.
-        crop_shape: The 3d center crop. Defaults to CROP_SHAPE_3D.
-
-    Returns:
-        The configuration dict.
-    """
-    is_em = dataset_name in DATASETS_3D_EM
-    is_3d = dataset_name in DATASETS_3D
-
-    if mode == "apg":
-        postproc_mode, grid = "apg", (APG_GRID_3D if is_3d else APG_GRID_2D)
-    else:
-        postproc_mode, grid = ("dense", EM_GRID) if is_em else ("sparse", LM_GRID)
-
-    return {
-        "mode": postproc_mode,
-        # Which metrics compute_metrics reports. Follows the data, not the mode, so that an APG run
-        # on EM is still ranked by the CREMI score.
-        "metric_mode": "dense" if is_em else "sparse",
-        "ndim": 3 if is_3d else 2,
-        "grid": grid,
-        "spacing": DATASET_SPACING.get(dataset_name, None),
-        "crop": (crop_shape or CROP_SHAPE_3D) if is_3d else None,
-        "criterion": criterion or ("cremi" if is_em else "msa"),
-        # Only set for the volumes with no split of their own.
-        "z_range": VAL_Z_RANGE.get(dataset_name),
-    }
-
-
-def compute_metrics(seg, labels, metric_mode, border_min_size=0):
-    """Compute the evaluation metrics for one segmentation.
-
-    Always reports mSA. For dense (EM neuron) mode it additionally reports the CREMI score and its
-    VI-split / VI-merge / adapted-Rand components, since neuron segmentation is ranked by CREMI
-    (lower is better) rather than mSA.
-    """
-    if seg.ndim == 2:
-        # Symmetric with the ground truth, which `load_evaluation_sample_2d` filtered the same way.
-        seg = drop_severed_objects(seg, border_min_size)
-    metrics = {"msa": float(mean_segmentation_accuracy(seg, labels))}
-    if metric_mode == "dense":
-        from elf.evaluation import cremi_score
-        vi_split, vi_merge, adapted_rand, cremi = cremi_score(seg, labels)
-        metrics["cremi"] = float(cremi)
-        metrics["vi_split"] = float(vi_split)
-        metrics["vi_merge"] = float(vi_merge)
-        metrics["adapted_rand"] = float(adapted_rand)
-    return metrics
-
-
-def deduplicate_flow_travel(params_list):
-    """Drop sparse combos whose flow travel duplicates a cheaper one.
-
-    'n_iter' and 'dt' act on the segmentation only through their product, the distance a pixel is
-    advected. Keeping the smallest 'n_iter' per product therefore covers the same travel distances
-    with fewer combos and fewer integration steps. The equivalence is empirical rather than exact,
-    since the integrator's discretization error still depends on 'dt'.
-    """
-    best, order = {}, []
-    for params in params_list:
-        if "n_iter" not in params or "dt" not in params:
-            return params_list
-        others = tuple(sorted((k, v) for k, v in params.items() if k not in ("n_iter", "dt")))
-        key = (others, round(params["n_iter"] * params["dt"], 6))
-        if key not in best:
-            best[key] = params
-            order.append(key)
-        elif params["n_iter"] < best[key]["n_iter"]:
-            best[key] = params
-    return [best[key] for key in order]
-
-
-def score_image_sparse_cached(
-    prediction, labels, params_list, backend, n_threads=POSTPROC_THREADS, spacing=None, border_min_size=0,
-):
-    """Score all sparse combos on one image, caching the flow density across shared combos.
-
-    Faithfully reproduces flow_instance_segmentation: the height map and each convergence-density map
-    (the expensive flow integration) are computed once and reused; only the cheap seed / watershed /
-    min-size steps vary per (density_threshold, min_size). Returns a per-combo list of metric dicts
-    (None where a combo failed), aligned with params_list.
-    """
-    foreground = prediction[0]
-    directed = prediction[1:]
-    ndim = foreground.ndim
-    if directed.shape[0] > ndim:
-        directed = directed[-ndim:]
-
-    # The convergence densities and the height maps are built up front, so the scoring below only reads them.
-    fg_mask_cache, density_cache, hmap_cache = {}, {}, {}
-    for params in params_list:
-        ft, sigma, n_iter, dt = (params[k] for k in FLOW_DENSITY_KEYS)
-        if ft not in fg_mask_cache:
-            fg_mask_cache[ft] = foreground > ft
-        key = (ft, sigma, n_iter, dt)
-        if key not in density_cache:
-            density_cache[key] = _compute_flow_density(
-                directed, fg_mask_cache[ft], n_iter=int(n_iter), dt=dt, sigma=sigma, spacing=spacing,
-                backend=backend, n_threads=n_threads,
-            )
-        fw = params["foreground_weight"]
-        if fw not in hmap_cache:
-            hmap_cache[fw] = watershed_heightmap(foreground, directed, fw)
-
-    # The base watershed does not depend on min_size, so all min_size values of a combo reuse it.
-    base_cache, base_lock = {}, threading.Lock()
-
-    def base_segmentation(key, fg_mask, density, density_threshold, hmap):
-        with base_lock:
-            cached = base_cache.get(key)
-        if cached is None:
-            seeds = connected_components(density > density_threshold)
-            cached = watershed(hmap, markers=seeds, mask=fg_mask)
-            with base_lock:
-                base_cache[key] = cached
-        return cached
-
-    def score(params):
-        ft, sigma, n_iter, dt = (params[k] for k in FLOW_DENSITY_KEYS)
-        fw, density_threshold = params["foreground_weight"], params["density_threshold"]
-        fg_mask = fg_mask_cache[ft]
-        hmap = hmap_cache[fw]
-        try:
-            key = (ft, sigma, n_iter, dt, density_threshold, fw)
-            seg = base_segmentation(key, fg_mask, density_cache[(ft, sigma, n_iter, dt)], density_threshold, hmap)
-            min_size = params["min_size"]
-            if min_size > 0:
-                seg = seg.copy()
-                ids, sizes = np.unique(seg, return_counts=True)
-                discard = ids[(sizes < min_size) & (ids > 0)]
-                seg[np.isin(seg, discard)] = 0
-                seg = watershed(hmap, markers=seg, mask=fg_mask)
-            return compute_metrics(seg.astype("uint32"), labels, "sparse", border_min_size)
-        except Exception as e:
-            warnings.warn(f"Sparse postprocessing failed for {params}: {e}")
-            return None
-
-    # Scoring a combination is independent of the others, and dominates the runtime of the sweep.
-    with futures.ThreadPoolExecutor(n_threads) as tp:
-        return list(tp.map(score, params_list))
-
-
-def dense_boundary_and_distances(prediction):
-    """Return the (boundary_map, in-plane distances) inputs to run_multicut for a dense prediction."""
-    fg = prediction[0]
-    boundary_map = fg.max() - fg
-    denom = boundary_map.max()
-    if denom > 0:
-        boundary_map = boundary_map / denom
-    distances = np.stack([prediction[2], prediction[3]])
-    return boundary_map, distances
-
-
-def dense_oversegmentation(boundary_map, distances, density_threshold, sigma, n_iter, dt, backend, n_threads):
-    """The expensive, beta-independent part of run_multicut: oversegmentation, RAG and edge features."""
-    from elf.segmentation.features import compute_rag, compute_boundary_mean_and_length, compute_z_edge_mask
-    n_slices = boundary_map.shape[0]
-    overseg = np.zeros(boundary_map.shape, dtype="uint64")
-
-    def run_overseg(z):
-        bd = boundary_map[z]
-        dists = distances[:, z]
-        fg_mask = np.ones(bd.shape, dtype="bool")
-        density = _compute_flow_density(
-            dists, fg_mask, n_iter=int(n_iter), dt=dt, sigma=sigma, verbose=False, backend=backend, n_threads=1,
-        )
-        seeds = connected_components(density > density_threshold)
-        bd = bd if np.issubdtype(bd.dtype, np.floating) else bd.astype("float32")
-        wsz = watershed(bd, markers=seeds)
-        overseg[z] = wsz
-        return int(wsz.max())
-
-    with futures.ThreadPoolExecutor(n_threads) as tp:
-        offsets = list(tp.map(run_overseg, range(n_slices)))
-
-    offsets = np.array(offsets, dtype="uint64")
-    offsets = np.roll(offsets, 1)
-    offsets[0] = 0
-    overseg += np.cumsum(offsets)[:, None, None]
-
-    rag = compute_rag(overseg)
-    feats = compute_boundary_mean_and_length(rag, overseg, boundary_map)
-    z_edges = None if n_slices == 1 else compute_z_edge_mask(rag, overseg)
-    return overseg, rag, feats, z_edges, n_slices
-
-
-def dense_solve(rag, feats, z_edges, overseg, n_slices, beta):
-    """Run the beta-dependent edge costs + multicut solve on a cached oversegmentation."""
-    from elf.segmentation.features import project_node_labels_to_pixels
-    from elf.segmentation.multicut import compute_edge_costs, multicut_decomposition
-    if n_slices == 1:
-        costs = compute_edge_costs(feats[:, 0], edge_sizes=feats[:, 1], weighting_scheme="all", beta=beta)
-    else:
-        costs = compute_edge_costs(
-            feats[:, 0], edge_sizes=feats[:, 1], weighting_scheme="xyz", z_edge_mask=z_edges, beta=beta,
-        )
-    node_labels = multicut_decomposition(rag, costs)
-    seg = project_node_labels_to_pixels(rag, overseg, node_labels)
-    return seg.astype("uint32")
-
-
-def score_image_dense_cached(prediction, labels, params_list, backend, n_threads=POSTPROC_THREADS, border_min_size=0):
-    """Score all dense combos on one image, caching the oversegmentation and RAG across shared combos."""
-    boundary_map, distances = dense_boundary_and_distances(prediction)
-    overseg_cache = {}
-    scores = []
-    for params in params_list:
-        dt = params.get("dt", 0.5)
-        key = tuple(params.get(k, dt if k == "dt" else None) for k in OVERSEG_KEYS)
-        if key not in overseg_cache:
-            overseg_cache[key] = dense_oversegmentation(
-                boundary_map, distances, params["density_threshold"], params["sigma"], params["n_iter"],
-                dt=dt, backend=backend, n_threads=n_threads,
-            )
-        overseg, rag, feats, z_edges, n_slices = overseg_cache[key]
-        try:
-            seg = dense_solve(rag, feats, z_edges, overseg, n_slices, params["beta"])
-            scores.append(compute_metrics(seg, labels, "dense", border_min_size))
-        except Exception as e:
-            warnings.warn(f"Dense multicut solve failed for {params}: {e}")
-            scores.append(None)
-    return scores
-
-
-def score_sample_apg(segmenter, raw, labels, params_list, ndim, metric_mode, spacing=None, border_min_size=0):
-    """Return a per-combo list of metric dicts for one sample, aligned with params_list.
-
-    The encoder and the decoder run once per sample. An image then reuses its mask proposals across
-    every combo that only selects among them; a volume gates its propagation on the score, so it has
-    to run every combo in full.
-    """
-    segmenter.clear_state()
-    # A volume takes the options that stop it re-reading its features on every propagation pass.
-    segmenter.initialize(raw, ndim=ndim, **(VOLUME_SPEED_OPTIONS if ndim == 3 else {}))
-    if ndim == 2:
-        return score_sample_apg_cached(segmenter, labels, params_list, metric_mode, border_min_size)
-
-    scores = []
-    for params in params_list:
-        try:
-            seg = segmenter.generate(spacing=spacing, **params)
-            scores.append(compute_metrics(seg.astype("uint32"), labels, metric_mode, border_min_size))
-        except Exception as e:
-            warnings.warn(f"Prompt generation failed for {params}: {e}")
-            scores.append(None)
-    return scores
-
-
-def score_sample_apg_cached(segmenter, labels, params_list, metric_mode, border_min_size=0):
-    """Score every combo of an image, prompting SAM2 once per distinct set of proposal parameters."""
-    groups = {}
-    for index, params in enumerate(params_list):
-        key = tuple(sorted((k, v) for k, v in params.items() if k in APG_PROPOSAL_KEYS))
-        groups.setdefault(key, []).append(index)
-
-    scores = [None] * len(params_list)
-    for key, indices in groups.items():
-        try:
-            proposals = segmenter.propose(**dict(key))
-        except Exception as e:
-            warnings.warn(f"Prompt generation failed for {dict(key)}: {e}")
-            continue
-        for index in indices:
-            selection = {k: v for k, v in params_list[index].items() if k not in APG_PROPOSAL_KEYS}
-            try:
-                seg = segmenter.select(proposals, **selection)
-                scores[index] = compute_metrics(seg.astype("uint32"), labels, metric_mode, border_min_size)
-            except Exception as e:
-                warnings.warn(f"Selection failed for {params_list[index]}: {e}")
-    return scores
-
-
-def resolve_checkpoint_identity(mode, model_type, joint_checkpoint="best", checkpoint_path=None):
-    """Return the content identity of the effective weights and, if needed, the joint checkpoint."""
-    joint_checksum = None
-    if checkpoint_path is None or mode == "apg":
-        joint_path = get_joint_checkpoint(model_type, joint_checkpoint)
-        joint_checksum = checkpoint_checksum(joint_path)
-
-    if checkpoint_path is None:
-        return joint_checksum, joint_checksum
-
-    decoder_checksum = checkpoint_checksum(checkpoint_path)
-    if mode == "apg":
-        return combine_checkpoint_checksums(joint_checksum, decoder_checksum), joint_checksum
-    return decoder_checksum, None
-
-
-def build_model(
-    mode, model_type, device, ndim, joint_checkpoint="best", checkpoint_path=None, joint_checksum=None,
-):
-    """Load the model a mode runs on, from the two halves of the joint checkpoint.
-
-    Args:
-        mode: The segmentation mode, one of MODES.
-        model_type: The SAM2 backbone of the joint model, e.g. 'hvit_t'.
-        device: The torch device.
-        ndim: The number of spatial dimensions, 2 or 3.
-        joint_checkpoint: The joint trainer checkpoint, without the '.pt' suffix.
-        checkpoint_path: Decoder weights to use instead of the ones exported from the joint checkpoint.
-
-    Returns:
-        The UniSAM2 decoder for 'ais', or the prompt generator for 'apg'.
-    """
-    if mode == "apg":
-        return build_apg_segmenter(
-            model_type, ndim, device, joint_checkpoint, decoder_path=checkpoint_path,
-            joint_checksum=joint_checksum,
-        )
-
-    decoder_path = checkpoint_path or export_joint_checkpoint(
-        model_type, joint_checkpoint, source_checksum=joint_checksum
-    )[1]
-    return load_unisam2_model(decoder_path, device, encoder=model_type)
 
 
 def segment(model, mode, raw, ndim, dataset_name, params, device, backend="cpp", spacing=None):
@@ -460,155 +46,6 @@ def segment(model, mode, raw, ndim, dataset_name, params, device, backend="cpp",
     return postprocess_unisam2(prediction, dataset_name, backend=backend, params=params)
 
 
-def tune_parameters(
-    model, mode, dataset_name, data_root, model_type, output_root, device,
-    backend="cpp", n_threads=POSTPROC_THREADS, n_tuning_samples=None, crop_shape=None, criterion=None,
-    checkpoint_id=None,
-):
-    """Sweep the grid of a mode on the validation split and return the best parameter combination.
-
-    The sweep and the evaluation load their samples through the same `common.load_data`, so the
-    parameters are selected on data that is preprocessed exactly like the data they are scored on.
-    Inference runs once per sample and every combination is scored on it before moving on, so the
-    predictions are never all held in memory at once.
-
-    The result is written below '<output_root>/<model_type>/<checkpoint_id>', which is the layout
-    `common.read_tuned_params` reads. An existing file is loaded rather than recomputed, so a
-    preempted job resumes at the evaluation instead of sweeping again.
-
-    Args:
-        model: The model of the mode, from `build_model`.
-        mode: The segmentation mode, 'ais' or 'apg'.
-        dataset_name: The dataset to tune on.
-        data_root: The root the data lives in.
-        model_type: The SAM2 backbone, which names the output subdirectory.
-        output_root: The root the sweep results are written to.
-        device: The torch device.
-        backend: The backend for the flow computation.
-        n_threads: The threads for the postprocessing.
-        n_tuning_samples: Cap the sweep to this many validation samples.
-        crop_shape: The 3d center crop.
-        criterion: The metric the grid is ranked by.
-        checkpoint_id: The checksum of all model weights used by the mode.
-
-    Returns:
-        The best parameter combination, or None when the dataset has nothing held out to tune on.
-    """
-    if not has_val_split(dataset_name):
-        print(f"'{dataset_name}' has no data held out from the evaluation, so the defaults are used.")
-        return None
-
-    config = tuning_config(dataset_name, mode, criterion=criterion, crop_shape=crop_shape)
-    output_dir = os.path.join(output_root, model_type, checkpoint_id) if checkpoint_id else os.path.join(
-        output_root, model_type
-    )
-    csv_path = os.path.join(output_dir, f"{dataset_name}.csv")
-    os.makedirs(output_dir, exist_ok=True)
-
-    legacy_path = os.path.join(output_root, model_type, f"{dataset_name}.csv")
-    if os.path.exists(csv_path) or (checkpoint_id is not None and os.path.exists(legacy_path)):
-        params = read_tuned_params(output_root, dataset_name, model_type, checkpoint_id)
-        cached_path = csv_path if os.path.exists(csv_path) else legacy_path
-        print(f"Loading the finished sweep at '{cached_path}'.")
-        report_best(pd.read_csv(cached_path), dataset_name, config)
-        return params
-
-    ndim, postproc_mode = config["ndim"], config["mode"]
-    keys = list(config["grid"])
-    params_list = [dict(zip(keys, combo)) for combo in itertools.product(*[config["grid"][k] for k in keys])]
-    if postproc_mode == "sparse":
-        n_full = len(params_list)
-        params_list = deduplicate_flow_travel(params_list)
-        if len(params_list) < n_full:
-            print(f"Deduplicated {n_full} combos to {len(params_list)} by flow travel (n_iter * dt).")
-
-    border_min_size = GT_MIN_SIZE_2D.get(dataset_name, 0) if ndim == 2 else 0
-    total = n_samples(dataset_name, data_root, split="val")
-    if n_tuning_samples is not None:
-        total = min(total, n_tuning_samples)
-    print(f"{dataset_name}: {len(params_list)} combinations over {total} validation sample(s), mode='{mode}'.")
-
-    samples = load_data(
-        dataset_name, data_root, ndim, split="val", crop_shape=config["crop"], z_range=config["z_range"],
-    )
-    metric_lists = [[] for _ in params_list]
-    t0 = time.perf_counter()
-    for index, (raw, labels, _) in enumerate(tqdm(samples, total=total, desc=f"tune-{mode}")):
-        if n_tuning_samples is not None and index >= n_tuning_samples:
-            break
-        if labels.max() == 0:  # Nothing to score without ground-truth.
-            continue
-        try:
-            if postproc_mode == "apg":
-                scores = score_sample_apg(
-                    model, raw, labels, params_list, ndim, config["metric_mode"],
-                    spacing=config["spacing"], border_min_size=border_min_size,
-                )
-            else:
-                prediction = predict_unisam2(model, raw, ndim=ndim, device=device)
-                scores = score_image(
-                    prediction, labels, postproc_mode, params_list, backend,
-                    n_threads=n_threads, spacing=config["spacing"], border_min_size=border_min_size,
-                )
-        except Exception as e:
-            warnings.warn(f"Skipping validation sample {index} of '{dataset_name}': {e}")
-            continue
-        for i, metrics in enumerate(scores):
-            if metrics is not None:
-                metric_lists[i].append(metrics)
-
-    rows = []
-    for params, per_sample in zip(params_list, metric_lists):
-        if not per_sample:
-            continue
-        row = {**params, "n_images": len(per_sample)}
-        for metric_key in per_sample[0]:
-            values = np.asarray([m[metric_key] for m in per_sample], dtype="float64")
-            row[f"{metric_key}_mean"] = float(values.mean())
-            row[f"{metric_key}_std"] = float(values.std())
-        rows.append(row)
-
-    if not rows:
-        warnings.warn(f"The sweep for '{dataset_name}' scored nothing, falling back to the defaults.")
-        return None
-
-    ascending = CRITERION_ASCENDING[config["criterion"]]
-    df = pd.DataFrame(rows).sort_values(f"{config['criterion']}_mean", ascending=ascending).reset_index(drop=True)
-    df.to_csv(csv_path, index=False)
-    print(f"Saved {csv_path} ({time.perf_counter() - t0:.0f}s).")
-    report_best(df, dataset_name, config)
-    return read_tuned_params(output_root, dataset_name, model_type, checkpoint_id)
-
-
-def score_image(
-    prediction, labels, mode, params_list, backend, n_threads=POSTPROC_THREADS, spacing=None, border_min_size=0,
-):
-    """Return a per-combo list of metric dicts for one decoder prediction, aligned with params_list."""
-    if mode == "sparse":
-        return score_image_sparse_cached(
-            prediction, labels, params_list, backend, n_threads=n_threads, spacing=spacing,
-            border_min_size=border_min_size,
-        )
-    return score_image_dense_cached(
-        prediction, labels, params_list, backend, n_threads=n_threads, border_min_size=border_min_size
-    )
-
-
-def report_best(df, dataset_name, config):
-    """Print the best parameter combination of a finished sweep.
-
-    The values are read back from the CSV for the report only. `common.read_tuned_params` is what
-    turns them back into the types `generate` and the postprocessing take.
-    """
-    criterion = config["criterion"]
-    ascending = CRITERION_ASCENDING[criterion]
-    best = df.sort_values(f"{criterion}_mean", ascending=ascending).iloc[0]
-    params = {key: best[key] for key in config["grid"]}
-    direction = "lower is better" if ascending else "higher is better"
-    print(f"Best params for {dataset_name} (by {criterion}, {direction}): {params}")
-    print(f"{criterion} = {best[f'{criterion}_mean']:.4f} (+/-{best[f'{criterion}_std']:.4f})")
-
-
 def run_evaluation(
     model, mode, dataset_name, data_root, experiment_folder, model_type, params, device,
     backend="cpp", crop_shape=None, checkpoint_id=None,
@@ -619,7 +56,7 @@ def run_evaluation(
     to the run that produced it.
 
     Args:
-        model: The model of the mode, from `build_model`.
+        model: The model of the mode, from `common.build_model`.
         mode: The segmentation mode, one of MODES.
         dataset_name: The dataset to score.
         data_root: The root the data lives in.
@@ -694,11 +131,8 @@ def main():
     parser.add_argument("-c", "--checkpoint", type=str, default=None, help="Weights instead of the joint export.")
     parser.add_argument("--joint_checkpoint", type=str, default="best", help="Joint checkpoint name, without '.pt'.")
     parser.add_argument("--skip_tuning", action="store_true", help="Evaluate with the library defaults.")
-    parser.add_argument("--tuning_root", type=str, default=None, help="Where the sweeps are stored and read from.")
-    parser.add_argument("--n_tuning_samples", type=int, default=None, help="Cap the sweep to this many samples.")
-    parser.add_argument("--criterion", type=str, default=None, choices=sorted(CRITERION_ASCENDING))
+    parser.add_argument("--tuning_root", type=str, default=None, help="Where parameter_search.py wrote its sweeps.")
     parser.add_argument("--backend", type=str, default="cpp", choices=("cpp", "python"), help="Flow backend.")
-    parser.add_argument("--n_threads", type=int, default=POSTPROC_THREADS, help="Threads for the postprocessing.")
     parser.add_argument("--crop_3d", type=int, nargs=3, default=None, help="Override the 3d crop (Z Y X).")
     args = parser.parse_args()
 
@@ -720,12 +154,18 @@ def main():
 
     params = None
     if not args.skip_tuning:
-        tuning_root = args.tuning_root or os.path.join(args.experiment_folder, "tuning", args.mode)
-        params = tune_parameters(
-            model, args.mode, args.dataset_name, args.input_path, args.model_type, tuning_root, device,
-            backend=args.backend, n_threads=args.n_threads, n_tuning_samples=args.n_tuning_samples,
-            crop_shape=crop_shape, criterion=args.criterion, checkpoint_id=checkpoint_id,
-        )
+        if has_val_split(args.dataset_name):
+            tuning_root = args.tuning_root or os.path.join(args.experiment_folder, "tuning")
+            tuning_root = os.path.join(tuning_root, args.mode)
+            try:
+                params = read_tuned_params(tuning_root, args.dataset_name, args.model_type, checkpoint_id)
+            except FileNotFoundError:
+                warnings.warn(
+                    f"No tuned parameters for '{args.dataset_name}' under '{tuning_root}'. Run "
+                    "parameter_search.py first, or pass --skip_tuning. Using the library defaults."
+                )
+        else:
+            print(f"'{args.dataset_name}' has no data held out from the evaluation, so the defaults are used.")
 
     run_evaluation(
         model, args.mode, args.dataset_name, args.input_path, args.experiment_folder, args.model_type,
